dynamic host bug fix
