+++ Dockerfile
@@ -20,6 +20,7 @@ RUN    ln -s /ibox/host-chroot.sh /ibox/blkid \
     && ln -s /ibox/host-chroot.sh /ibox/mount \
     # && ln -s /ibox/host-chroot.sh /ibox/multipath \
     # && ln -s /ibox/host-chroot.sh /ibox/multipathd \
+    # && ln -s /ibox/host-chroot.sh /ibox/stat \ 
     && ln -s /ibox/host-chroot.sh /ibox/cat \
     && ln -s /ibox/host-chroot.sh /ibox/mkdir \
     && ln -s /ibox/host-chroot.sh /ibox/rmdir \
